pyproject.toml
[tool.poetry]
name = "local-dev-env"
version = "0.1.0"
description = ""
authors = ["pacho-h <pacho.h90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
kafka-python = "^2.0.2"


[tool.poetry.group.dev.dependencies]
pynvim = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
